+++ src/neural_network/optimizer/Softmax.cpp
@@ -36,7 +36,7 @@ inline void Softmax::computeSoftmax(std::vector<float>& outputs)
                                         });
     for (auto& output : outputs)
     {
-        const auto value = exp(output) / sumExp;
+        const auto value = std::expf(output) / sumExp;
         if (std::fpclassify(value) != FP_NORMAL && std::fpclassify(value) != FP_ZERO)
         {
             output = 0;
